Be a ULong

+++ ksrpc-sockets/src/macosX64Main/kotlin/TermiosUtil.macosX64.kt
@@ -4,6 +4,6 @@ import platform.posix.ICANON
 import platform.posix.termios
 
 actual fun termios.setICanon() {
-    c_lflag = c_lflag and ICANON.inv().toUInt()
+    c_lflag = c_lflag and ICANON.inv().toULong()
 }
